+++ src/rustup.rs
@@ -4,7 +4,34 @@ use Target;
 use errors::*;
 use extensions::CommandExt;
 
-pub fn install(target: Target, verbose: bool) -> Result<()> {
+pub struct AvailableTargets {
+    triples: Vec<String>,
+}
+
+impl AvailableTargets {
+    pub fn contains(&self, target: &Target) -> bool {
+        let target = target.triple();
+        self.triples.iter().any(|t| t == target)
+    }
+}
+
+pub fn available_targets(verbose: bool) -> Result<AvailableTargets> {
+    let out = Command::new("rustup").args(&["target", "list"])
+        .run_and_get_stdout(verbose)?;
+
+    Ok(AvailableTargets {
+        triples: out.lines()
+            .filter_map(|line| if line.contains("installed") ||
+                                  line.contains("default") {
+                None
+            } else {
+                Some(line.to_owned())
+            })
+            .collect(),
+    })
+}
+
+pub fn install(target: &Target, verbose: bool) -> Result<()> {
     let target = target.triple();
 
     Command::new("rustup")
@@ -18,23 +45,6 @@ pub fn install_rust_src(verbose: bool) -> Result<()> {
         .args(&["component", "add", "rust-src"])
         .run(verbose)
         .chain_err(|| format!("couldn't install the `rust-src` component"))
-}
-
-pub fn available_targets(verbose: bool) -> Result<Vec<Target>> {
-    let out = Command::new("rustup").args(&["target", "list"])
-        .run_and_get_stdout(verbose)?;
-
-    Ok(out.lines()
-        .filter_map(|line| if line.contains("installed") ||
-                              line.contains("default") {
-            None
-        } else {
-            match Target::from(line) {
-                Target::Other => None,
-                t => Some(t),
-            }
-        })
-        .collect())
 }
 
 pub fn rust_src_is_installed(verbose: bool) -> Result<bool> {